Fixed SELinux relabelling after Swift reset
Swift reset makes a backup of the nodes before the tests start and copies them
back after the tests. However, this makes it lose SELinux labels (see ls -Z)
and so rsync fails because of AVC denials. Thanks to this patch, reset now runs
restorecon on /srv/node after the backup was restored.