Added extra parameters to Swift replica checks
Added 'exact' param, so that it can be tested if Swift has
exactly 'count' number of replicas, no more, no less.

Added 'check_nodes' to be able to specify that all the primarily nodes have a
copy of a file (to do so, set check_nodes=replica_count) and other uses.