Test module configs get generated from config.json
With 'bin/generate_config_files.py', each test module will get its own
configuration, created based on the 'etc/config.json' file, in which there will
be (for example) specified which servers are to be used as swift proxy servers
and which as data servers.

Added bin/ folder with additional (so far) empty scripts for OpenStack
deployment and state reset.

The OpenStack deployment is now configurable, so that a user can choose his own
installing method. This feature is still work in progress.